--- CsvImporter/src/main/java/devminds/example/JsonObject.java
@@ -1,104 +1,0 @@
-package devminds.example;
-
-public class JsonObject {
-    private String timestamp;
-    private String email;
-    private String emailFatec;
-    private String nomeCompleto;
-    private String nomeCompletoOrientador;
-    private String emailOrientador;
-    private String matriculadoEm;
-    private String tipoTG;
-    private String problema;
-    private String empresa;
-    private String disciplina;
-
-
-    public String getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(String timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getEmailFatec() {
-        return emailFatec;
-    }
-
-    public void setEmailFatec(String emailFatec) {
-        this.emailFatec = emailFatec;
-    }
-
-    public String getNomeCompleto() {
-        return nomeCompleto;
-    }
-
-    public void setNomeCompleto(String nomeCompleto) {
-        this.nomeCompleto = nomeCompleto;
-    }
-
-    public String getNomeCompletoOrientador() {
-        return nomeCompletoOrientador;
-    }
-
-    public void setNomeCompletoOrientador(String nomeCompletoOrientador) {
-        this.nomeCompletoOrientador = nomeCompletoOrientador;
-    }
-
-    public String getEmailOrientador() {
-        return emailOrientador;
-    }
-
-    public void setEmailOrientador(String emailOrientador) {
-        this.emailOrientador = emailOrientador;
-    }
-
-    public String getMatriculadoEm() {
-        return matriculadoEm;
-    }
-
-    public void setMatriculadoEm(String matriculadoEm) {
-        this.matriculadoEm = matriculadoEm;
-    }
-
-    public String getTipoTG() {
-        return tipoTG;
-    }
-
-    public void setTipoTG(String tipoTG) {
-        this.tipoTG = tipoTG;
-    }
-
-    public String getProblema() {
-        return problema;
-    }
-
-    public void setProblema(String problema) {
-        this.problema = problema;
-    }
-
-    public String getEmpresa() {
-        return empresa;
-    }
-
-    public void setEmpresa(String empresa) {
-        this.empresa = empresa;
-    }
-
-    public String getDisciplina() {
-        return disciplina;
-    }
-
-    public void setDisciplina(String disciplina) {
-        this.disciplina = disciplina;
-    }
-}
