+++ CsvImporter/src/main/java/devminds/example/Main.java
@@ -1,5 +1,7 @@
 package devminds.example;
 
+import devminds.example.csv.CsvImporter;
+
 public class Main {
     public static void main(String[] args) {
         CsvImporter csvImporter = new CsvImporter();
