+++ CsvImporter/src/main/java/devminds/example/csv/CsvReader.java
@@ -1,4 +1,6 @@
-package devminds.example;
+package devminds.example.csv;
+import devminds.example.json.JsonFileMaker;
+import devminds.example.json.JsonObject;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -8,6 +10,7 @@ public class CsvReader {
         try {
             FileReader fileReader = new FileReader(filename);
             BufferedReader br = new BufferedReader(fileReader);
+            br.readLine();
             String line = "";
             String[] nextRecord;
             int indexFor = 0;
@@ -16,13 +19,13 @@ public class CsvReader {
                 JsonObject jsonOBJ = new JsonObject();
                 indexFor = 0;
                 for (String tempStr : nextRecord) {
-                    System.out.println(indexFor + " " + tempStr + " ");
+//                    System.out.println(indexFor + " " + tempStr + " ");
                     try{
                     switch (indexFor) {
                         case 0:
                             jsonOBJ.setTimestamp(nextRecord[0]);
                         case 1:
-                            jsonOBJ.setEmail(nextRecord[1]);
+
                         case 2:
                             jsonOBJ.setEmailFatec(nextRecord[2]);
                         case 3:
@@ -32,7 +35,7 @@ public class CsvReader {
                         case 5:
                             jsonOBJ.setEmailOrientador(nextRecord[5]);
                         case 6:
-                            jsonOBJ.setMatriculadoEm(nextRecord[6]);
+
                         case 7:
                             jsonOBJ.setTipoTG(nextRecord[7]);
                         case 8:
@@ -44,13 +47,12 @@ public class CsvReader {
                         default:
 
                     }
-                    } catch (Exception e){
-                        System.out.println(e);
-                    }
-                    JsonFileMaker fileMaker = new JsonFileMaker();
-                    fileMaker.salvaJson(jsonOBJ);
-                    indexFor += 1;
+
+                    } catch (Exception e){}
                 }
+                JsonFileMaker fileMaker = new JsonFileMaker();
+                fileMaker.salvaJson(jsonOBJ);
+                indexFor += 1;
 
 
             }
