+++ CsvImporter/src/main/java/devminds/example/json/JsonFileMaker.java
@@ -1,4 +1,4 @@
-package devminds.example;
+package devminds.example.json;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -11,7 +11,9 @@ public class JsonFileMaker {
     public void salvaJson(JsonObject object) throws IOException {
         final File dir = new File("jsonFilesOutput");
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        FileWriter escritor = new FileWriter(new File(dir,object.getNomeCompleto() + ".json"));
+        String timestampTemp = object.getTimestamp();
+        String aux = timestampTemp.replaceAll("[^a-zA-Z0-9]", "");
+        FileWriter escritor = new FileWriter(new File(dir,object.getNomeCompleto() + "_" + aux+ ".json"));
         escritor.write(gson.toJson(object));
         escritor.close();
     }
