+++ CsvImporter/src/main/java/devminds/example/objetos/Aluno.java
@@ -1,0 +1,23 @@
+package devminds.example.objetos;
+
+public class Aluno {
+    private String emailFatec;
+    private String nomeCompleto;
+
+
+    public String getEmailFatec() {
+        return emailFatec;
+    }
+
+    public void setEmailFatec(String emailFatec) {
+        this.emailFatec = emailFatec;
+    }
+
+    public String getNomeCompleto() {
+        return nomeCompleto;
+    }
+
+    public void setNomeCompleto(String nomeCompleto) {
+        this.nomeCompleto = nomeCompleto;
+    }
+}
